+++ rejoice/src/bin/commands/mod.rs
@@ -1,9 +1,12 @@
+mod boilerplate;
 mod build;
 mod dev;
 mod init;
 mod islands;
+mod migrate;
 pub mod style;
 
 pub use build::build_command;
 pub use dev::dev_command;
 pub use init::init_command;
+pub use migrate::{migrate_add, migrate_revert, migrate_status, migrate_up};